// src/app/main/shared/dtos/model/requestConsulta.ts
/**
 * API Recaudos y Desembolsos Sura - Colombia
 * Esta API ofrece operaciones para realizar recaudo, consulta de recaudo, creación de aviso (legalización) y registro de cartera. Sura - Colombia.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: sopinfo@sura.com.co
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { HeadRequestConsulta } from './headRequestConsulta';

/**
 * Parámetros para consulta de avisos
 */
export interface RequestConsulta { 
    headRequest?: HeadRequestConsulta;
    /**
     * Identificador del aviso, se obtiene del campo cdAviso del servicio de session
     */
    avsid: string;
    /**
     * Fuente asociada, es entregada al aliado y es diferente para ambiente de laboratorio y producción
     */
    cdFuente?: string;
    /**
     * Fecha inicio de búsqueda
     */
    fechaIni?: string;
    /**
     * Fecha fin de búsqueda
     */
    fechaFin?: string;
    /**
     * Estado del aviso, enviar siempre vacío
     */
    estado?: string;
    /**
     * Identificación del cliente
     */
    nmDocumento?: string;
    /**
     * Forma de pago, enviar siempre vacío
     */
    formaPago?: string;
}